# api/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PudimFinance API",
    "description": "Personal finance application API. Immutable double-entry ledger, event sourcing, real-time reporting, receipt scanning and price tracking.",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
  "paths": {
...
    "/api/sync/pull": {
      "post": {
        "tags": [
          "Sync"
        ],
        "summary": "Pulls entities changed since `last_synced_at`.",
        "operationId": "pull",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyncPullRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Changed entities since last sync",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncPullResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/sync/push": {
      "post": {
        "tags": [
          "Sync"
        ],
        "summary": "Applies a batch of client mutations.",
        "operationId": "push",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyncPushRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Per-operation results",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncPushResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Category": {
        "type": "object",
        "description": "A transaction category (income or expense), optionally nested via `parent_id`.",
        "required": [
          "id",
          "name",
          "type",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "color": {
            "type": [
              "string",
              "null"
            ],
            "description": "Hex color code (e.g., `#ef4444`)."
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Row creation timestamp."
          },
          "icon": {
            "type": [
              "string",
              "null"
            ],
            "description": "Icon identifier used by the web/mobile UIs."
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Unique category identifier."
          },
          "name": {
            "type": "string",
            "description": "Display name (e.g., \"Food & Groceries\")."
          },
          "parent_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid",
            "description": "Optional parent category for subcategories."
          },
          "type": {
            "type": "string",
            "description": "`income` or `expense`."
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "Last row update timestamp."
          }
        }
      },
...
      "SyncOpResult": {
        "type": "object",
        "description": "Result for a single pushed operation.",
        "required": [
          "client_id",
          "status"
        ],
        "properties": {
          "client_id": {
            "type": "string",
            "description": "Echo of the client_id."
          },
          "error": {
            "type": [
              "string",
              "null"
            ],
            "description": "Error message when status != ok."
          },
          "server_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid",
            "description": "Server-assigned UUID for creates."
          },
          "status": {
            "type": "string",
            "description": "`ok`, `conflict`, or `error`."
          }
        }
      },
      "SyncOperation": {
        "type": "object",
        "description": "A single client-initiated mutation to apply on the server.",
        "required": [
          "operation_type",
          "entity_type",
          "client_id",
          "payload"
        ],
        "properties": {
          "client_id": {
            "type": "string",
            "description": "Client-generated UUID used as an idempotency key."
          },
          "entity_type": {
            "type": "string",
            "description": "`transaction` or `category`."
          },
          "operation_type": {
            "type": "string",
            "description": "`create`, `update`, or `delete`."
          },
          "payload": {
            "description": "Request body (e.g. a CreateTransactionRequest) for create/update."
          },
          "server_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid",
            "description": "Server UUID for `update`/`delete` operations."
          }
        }
      },
      "SyncPullRequest": {
        "type": "object",
        "description": "Request: pull changes since a given timestamp.",
        "required": [
          "last_synced_at"
        ],
        "properties": {
          "last_synced_at": {
            "type": "string",
            "format": "date-time",
            "description": "Only return rows with `updated_at` after this timestamp."
          }
        }
      },
      "SyncPullResponse": {
        "type": "object",
        "description": "Response: entities changed since the client's last sync.",
        "required": [
          "categories",
          "transactions",
          "server_time"
        ],
        "properties": {
          "categories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Category"
            },
            "description": "All categories (or those changed since last sync)."
          },
          "server_time": {
            "type": "string",
            "format": "date-time",
            "description": "Server time — client stores this as its next `last_synced_at`."
          },
          "transactions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Transaction"
            },
            "description": "Transactions changed since last sync."
          }
        }
      },
      "SyncPushRequest": {
        "type": "object",
        "description": "Request: batch of client mutations.",
        "required": [
          "operations"
        ],
        "properties": {
          "operations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SyncOperation"
            },
            "description": "Operations to apply, in order."
          }
        }
      },
      "SyncPushResponse": {
        "type": "object",
        "description": "Response: results for each pushed operation.",
        "required": [
          "results"
        ],
        "properties": {
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SyncOpResult"
            },
            "description": "Per-operation results (same order as the request)."
          }
        }
      },
      "Transaction": {
        "type": "object",
        "description": "A single income or expense transaction.",
        "required": [
          "id",
          "description",
          "amount",
          "type",
          "date",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "account_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid",
            "description": "Source account (payment method, e.g. a credit card) used for this\ntransaction (NULL when unlinked)."
          },
          "amount": {
            "type": "string",
            "description": "Monetary amount — always positive; `type` determines direction."
          },
          "category_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid",
            "description": "Category this transaction belongs to (nullable if category deleted)."
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Row creation timestamp."
          },
          "date": {
            "type": "string",
            "format": "date",
            "description": "Calendar date of the transaction."
          },
          "description": {
            "type": "string",
            "description": "Human-readable description (e.g., \"Lunch at Restaurante X\")."
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Unique transaction identifier."
          },
          "installment_plan_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid",
            "description": "Installment plan this transaction belongs to (NULL for regular transactions)."
          },
          "notes": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional free-form notes."
          },
          "type": {
            "type": "string",
            "description": "`income` or `expense`."
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "Last row update timestamp."
          }
        }
      },
...
    }
  },
  "security": [
    {}
  ],
...
}